daemon: update annotations if CNP rule spec has not changed
Previously, if a user updated the annotations for a rule, but did not update the
\`spec\` or \`specs\` fields, the logic in the daemon would short circuit
updating the rule in the local repository within Cilium to avoid unneeded and
costly policy recalculation. This meant that the annotations would not be
updated. Add logic to update the annotations if they have changed even if the
rule's \`spec\` or \`specs\` have not changed. Also ensure that we create a deep
copy of the annotations when adding a rule into the Cilium repository.

Signed-off by: Ian Vernon <ian@cilium.io>"F