etcd: Fix and relax during recreate watchr loop
I can't think of a reason why the revision to watch should be bumped when
Watch() returns a closed channel. We have also observed occasions where etcd
returns a closed channel on calling Watch(). The problem typically resolves
itself quickly but we should sleep for a while to relax the CPU.

Signed-off-by: Thomas Graf <thomas@cilium.io>